Add support for pattern properties
Convert the patterns to actual RegExp objects and then render them as such.
This is especially useful for the perfectly valid empty pattern, which
otherwise would not appear.

Add a function, `propertySpecs`, which returns specs for handling both
properties and patternProperties, and then iterate over them both to produce
both the summary table and the more detailed info. Right now it creates a
separate table for patter properties, but it might be worth considering a
single table, since the rendering of the regex makes it pretty clear which
is which.

Add tests for pattern properties, too.